+++ test/protractor.js
@@ -1,4 +1,4 @@
 exports.config = {
   seleniumAddress: 'http://localhost:4444/wd/hub',
-  specs: ['spec.*.js']
+  specs: ['spec/*.js']
 };
